more refactoring

+++ src/build-infra.sh
@@ -18,11 +18,12 @@ Help()
    echo "e     EC2 name. Example: t2.micro"
    echo "v     EBS volume size for EC2. In GB. Example: 200"
    echo "t     Type of process: test or all. Options: [test, all]"
+   echo "l     Process location: local or remote. Options: [local, remote]"
    echo "h     Print this Help."
    echo
 }
 
-while getopts "w:s:p:e:v:t:h" flag;
+while getopts "w:s:p:e:v:t:l:h" flag;
 do
     case $flag in
         w) WORKUNIT=$OPTARG;;
@@ -31,6 +32,7 @@ do
         e) EC2=$OPTARG;;
         v) VOLUME_SIZE=$OPTARG;;
         t) TYPE=$OPTARG;;
+        l) LOCATION=$OPTARG;;
         h) Help
            exit;;
         \?) # Invalid option
@@ -43,6 +45,7 @@ cp -r terraform terraform-${WORKUNIT}-${TYPE}
 
 cd terraform-${WORKUNIT}-${TYPE}
 
+## Set TF VARS
 export TF_VAR_instance_type="${EC2}"
 export TF_VAR_volume_size="${VOLUME_SIZE}"
 if [[ $TYPE == "test" ]]; then
@@ -54,13 +57,18 @@ fi
 export TF_VAR_workunit=$WORKUNIT
 export TF_VAR_state=$STATE
 
-terraform init 
-
-terraform apply -auto-approve
-
-terraform validate
-
-terraform destroy -auto-approve
+## Process Location
+if [[ $LOCATION == "local" ]]; then
+    echo "Testing Locally..."
+    bash ../${TYPE}/builds/${TF_VAR_process_file} ${WORKUNIT} ${STATE} ${PROCESS} ${EC2} ${VOLUME_SIZE} ${TYPE}
+elif [[ $LOCATION == "remote" ]]; then
+    terraform init 
+    terraform apply -auto-approve
+    terraform validate
+    terraform destroy -auto-approve
+else
+    echo "A location where to process, 'local' or 'remote', must be set"
+fi
 
 cd ../
